progress (i have to fix fragments. cause as of now different stacks share the same fragments reference)

+++ Source/XyloInventoryUtil/Public/Inventory/Fragment/XIUCountFragment.h
@@ -16,8 +16,8 @@ class XYLOINVENTORYUTIL_API UXIUCountFragment : public UXIUItemFragment
 
 public:
 	UPROPERTY(EditAnywhere)
-	float MaxCount;
+	int MaxCount;
 	
 	UPROPERTY(EditAnywhere)
-	float Count;
+	int Count;
 };
